still can't get rid of duplicates when searching multiple times, but other stuff is working.
