+++ src/main/java/com/launchcode/violetSwap/models/UserService.java
@@ -1,0 +1,34 @@
+package com.launchcode.violetSwap.models;
+
+import com.launchcode.violetSwap.models.data.UserRepository;
+import jakarta.servlet.http.HttpSession;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.stereotype.Service;
+
+import java.util.Optional;
+
+@Service
+public class UserService {
+
+    @Autowired
+    private UserRepository userRepository;
+
+    private static final String userSessionKey = "user";
+
+    public User getUserFromSession(HttpSession session) {
+        Integer userId = (Integer) session.getAttribute(userSessionKey);
+        if (userId == null) {
+            return null;
+        }
+
+        Optional<User> user = userRepository.findById(userId);
+
+        if(user.isEmpty()) {
+            return null;
+        }
+
+        return user.get();
+    }
+
+
+}
