+++ src/main/resources/templates/search/varieties.html
@@ -8,15 +8,21 @@
 <form method="post" style="max-width:600px;">
     <div class="form-group">
         <label>Search for Variety :
-            <input class="form-control" type="text" name="varietySearch" />
+            <input class="form-control" type="text" name="searchTerm" />
         </label>
         <!-- Include CSRF token -->
         <input type="hidden" th:name="${_csrf.parameterName}" th:value="${_csrf.token}" />
         <button type="submit">Search</button>
     </div>
-
-
 </form>
+
+<!--<form th:action="@{/search/varieties/}" method="get">-->
+<!--    <label for="searchTerm">Browse by Variety :</label>-->
+<!--    <input type="text" id="searchTerm" name="searchTerm" />-->
+<!--    <button type="submit">Search</button>-->
+<!--</form>-->
+
+
 <!-- List of Varieties -->
 <div>
     <ul>
